2009-08-03  Jean Brefort  <jean.brefort@normalesup.org> 	* libs/gcp/atom.cc: enhanced bond drawing near symbols and clean out white 	rectangles. 	* libs/gcp/fragment.cc: fix horizontal position. 	* libs/gcp/fragment.h: use double for left bearing. 	* libs/gcu/transform3d.cc: add missing const. 	* libs/gcu/transform3d.h: ditto. 	* plugins/paint/wikipedia/wikipediatool.cc: start exporting png files again.

+++ libs/gcp/fragment.h
@@ -268,7 +268,7 @@ private:
 private:
 	FragmentAtom *m_Atom;
 	unsigned m_BeginAtom, m_EndAtom;
-	int m_lbearing;
+	double m_lbearing;
 	double m_CHeight;
 	bool m_Inversable;
